workspace/rust: 1.90 → 1.95
The update to `flake.lock` is a simple

    nix flake update rust-overlay

to be able to reach 1.95.

+++ crates/radicle/src/cob/cache.rs
@@ -290,7 +290,7 @@ fn bump(db: &sql::Connection) -> Result<usize, Error> {
 
     db.execute(format!("PRAGMA user_version = {new}"))?;
 
-    Ok(new as usize)
+    Ok(new)
 }
 
 /// Update a COB object in the cache.
