+++ crates/radicle/src/node/db.rs
@@ -220,7 +220,7 @@ pub fn bump(db: &sql::Connection) -> Result<usize, Error> {
 
     db.execute(format!("PRAGMA user_version = {new}"))?;
 
-    Ok(new as usize)
+    Ok(new)
 }
 
 /// Migrate the database to the latest schema.
